+++ src/babashka/impl/tasks.clj
@@ -67,8 +67,9 @@
       (apply log-info args)
       (handle-non-zero (pp/process* {:opts opts :cmd cmd :prev prev}) opts))))
 
-(defn clojure [cmd & args]
-  (let [[opts cmd args]
+(defn clojure [& args]
+  (let [[cmd & args] args
+        [opts cmd args]
         (if (map? cmd)
           [cmd (first args) (rest args)]
           [nil cmd args])